+++ src/main/java/gr/athenarc/imsi/visualfacts/util/ContainmentExaminer.java
@@ -1,5 +1,5 @@
 package gr.athenarc.imsi.visualfacts.util;
 
 public interface ContainmentExaminer {
-    boolean contains(float x, float y);
+    boolean contains(double x, double y);
 }
